zoo/FSRE-Depth/requirements.txt
asn1crypto==1.0.1
backcall==0.1.0
beautifulsoup4==4.8.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
conda==4.7.12
conda-build==3.18.9
conda-package-handling==1.6.0
cryptography==2.7
cycler==0.10.0
decorator==4.4.0
filelock==3.0.12
glob2==0.7
idna==2.8
imageio==2.9.0
ipython==7.8.0
ipython-genutils==0.2.0
jedi==0.15.1
Jinja2==2.10.3
kiwisolver==1.3.1
libarchive-c==2.8
lief==0.9.0
MarkupSafe==1.1.1
matplotlib==3.3.4
mkl-fft==1.0.14
mkl-random==1.1.0
mkl-service==2.3.0
networkx==2.5.1
numpy==1.17.2
olefile==0.46
opencv-python==4.5.3.56
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.0
pkginfo==1.5.0.1
prompt-toolkit==2.0.10
protobuf==3.18.1
psutil==5.6.3
ptyprocess==0.6.0
pycosat==0.6.3
pycparser==2.19
Pygments==2.4.2
pyOpenSSL==19.0.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.1.2
requests==2.22.0
ruamel-yaml==0.15.46
scikit-image==0.17.2
scipy==1.5.4
six==1.12.0
soupsieve==1.9.3
tensorboardX==2.4
tifffile==2020.9.3
torch==1.3.0
torchvision==0.4.1a0+d94043a
tqdm==4.32.1
traitlets==4.3.3
urllib3==1.24.2
wcwidth==0.1.7
